perfil

+++ GestordeTareas.UI/Controllers/TareaController.cs
@@ -19,6 +19,8 @@ namespace GestordeTareas.UI.Controllers
         private readonly PrioridadBL _prioridadBL;
         private readonly EstadoTareaBL _estadoTareaBL;
         private readonly ProyectoBL _proyectoBL;
+        private readonly UsuarioBL _usuarioBL;
+        private readonly ProyectoUsuarioBL _proyectoUsuarioBL;
 
         public TareaController()
         {
@@ -27,14 +29,18 @@ namespace GestordeTareas.UI.Controllers
             _prioridadBL = new PrioridadBL();
             _estadoTareaBL = new EstadoTareaBL();
             _proyectoBL = new ProyectoBL();
+            _usuarioBL = new UsuarioBL();
+            _proyectoUsuarioBL = new ProyectoUsuarioBL();
         }
 
         // GET: TareaController
         public async Task<ActionResult> Index(int proyectoId)
         {
-            //List<Tarea> Lista = await _tareaBL.GetAllAsync();
-
-            //return View(Lista);
+            if (!await VerificarAcceso(proyectoId))
+            {
+                TempData["ErrorMessage"] = "No tienes acceso a este proyecto.";
+                return RedirectToAction("Index", "Proyecto"); // Redirigir a la vista de proyectos
+            }
 
             // Aquí cargas las tareas asociadas al proyecto con el ID proporcionado
             var tareas = await _tareaBL.GetTareasByProyectoIdAsync(proyectoId);
@@ -73,11 +79,16 @@ namespace GestordeTareas.UI.Controllers
             return PartialView("Create", tarea);
         }
 
-        // POST: CategoriaController/Create
+        // POST: TareaController/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create(Tarea tarea, int idProyecto)
         {
+            if (!User.IsInRole("Administrador"))
+            {
+                TempData["ErrorMessage"] = "No tienes permisos para realizar cambios.";
+                return View(tarea);
+            }
             try
             {
                 // Asignar el ID del proyecto a la tarea
@@ -87,19 +98,12 @@ namespace GestordeTareas.UI.Controllers
                 tarea.IdEstadoTarea = estadoPendienteId;
 
                 int result = await _tareaBL.CreateAsync(tarea);
-                return RedirectToAction(nameof(Index), new { id = idProyecto });
+                return Json(new { success = true, message = "Tarea creada correctamente.", id = idProyecto });
+
             }
             catch (Exception ex)
             {
-                ViewBag.Error = ex.Message;
-
-                // Volver a cargar las listas desplegables u otros datos necesarios para la vista
-                await LoadDropDownListsAsync();
-
-                //  ViewBag.idProyecto = GetProyectoIdAsync(proyecto);
-                // Devolver la vista parcial "Create" con la tarea y el ID de proyecto
-                return PartialView("Create", new Tarea { IdProyecto = idProyecto });
-                // return PartialView("Create", tarea);
+                return Json(new { success = false, message = "Error al crear la tarea: " + ex.Message });
             }
         }
 
@@ -134,16 +138,19 @@ namespace GestordeTareas.UI.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit(int id, Tarea tarea)
         {
+            if (!User.IsInRole("Administrador"))
+            {
+                TempData["ErrorMessage"] = "No tienes permisos para realizar cambios.";
+                return View(tarea);
+            }
             try
             {
                 int result = await _tareaBL.UpdateAsync(tarea);
-                return RedirectToAction(nameof(Index), new { id = tarea.IdProyecto });
+                return Json(new { success = true, message = "Tarea editada correctamente.", id = tarea.IdProyecto });
             }
             catch (Exception ex)
             {
-                ViewBag.Error = ex.Message;
-                await LoadDropDownListsAsync();
-                return View(tarea);
+                return Json(new { success = false, message = $"Error al editar la tarea: {ex.Message}" });
             }
         }
 
@@ -160,15 +167,19 @@ namespace GestordeTareas.UI.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Delete(int id, Tarea tarea)
         {
+            if (!User.IsInRole("Administrador"))
+            {
+                TempData["ErrorMessage"] = "No tienes permisos para realizar cambios.";
+                return View(tarea);
+            }
             try
             {
                 await _tareaBL.DeleteAsync(tarea);
-                return RedirectToAction(nameof(Index), new { id = tarea.IdProyecto });
+                return Json(new { success = true, message = "Tarea eliminada correctamente.", id = tarea.IdProyecto });
             }
             catch (Exception ex)
             {
-                ViewBag.Error = ex.Message;
-                return View(tarea);
+                return Json(new { success = false, message = $"Error al eliminar la tarea: {ex.Message}" });
             }
         }
 
@@ -207,6 +218,31 @@ namespace GestordeTareas.UI.Controllers
         {
             public int IdTarea { get; set; }
             public int IdEstadoTarea { get; set; }
+        }
+
+        // MÉTODO PARA VERIFICAR EL ACCESO A LAS TAREAS
+        private async Task<bool> VerificarAcceso(int idProyecto)
+        {
+            var users = await _usuarioBL.SearchAsync(new Usuario { NombreUsuario = User.Identity.Name, Top_Aux = 1 });
+            var actualUser = users.FirstOrDefault();
+
+            if (actualUser == null)
+            {
+                return false; // Usuario no encontrado
+            }
+
+            int idUsuario = actualUser.Id;
+
+            // Verificar si el usuario es administrador
+            bool esAdmin = User.IsInRole("Administrador");
+            if (esAdmin)
+            {
+                return true; // Acceso permitido para administradores
+            }
+
+            // Verificar si el usuario está unido al proyecto
+            var usuariosUnidos = await _proyectoUsuarioBL.ObtenerUsuariosUnidosAsync(idProyecto);
+            return usuariosUnidos.Any(u => u.Id == idUsuario); // Devuelve true si está unido, false si no
         }
 
     }
